시간표기능 중단, 학점계산기 만드는중

+++ KU_Board/board/views.py
@@ -12,6 +12,10 @@ from django.urls import reverse
 
 
 def board(request, whatboard):
+
+    if request.user.is_anonymous:
+        return redirect('/')
+
     if whatboard == 'free':
         board_post = Post.objects.filter(category=1).order_by('-boardNum')
         categorykr = '자유게시판'
@@ -62,6 +66,10 @@ def board(request, whatboard):
 
 
 def writingpost(request, whatboard):
+
+    if request.user.is_anonymous:
+        return redirect('/')
+
     if whatboard == 'free':
         category = 1
         changedwhatboard = '자유게시판'
@@ -100,6 +108,10 @@ def writingpost(request, whatboard):
 
 
 def postdetail(request, whatboard, pk):
+
+    if request.user.is_anonymous:
+        return redirect('/')
+
     if whatboard == 'free':
         categoryNum = 1
     elif whatboard == 'secret':
@@ -146,12 +158,16 @@ def postdetail(request, whatboard, pk):
 
 
 def recommend(request, whatboard, pk):
+
+    if request.user.is_anonymous:
+        return redirect('/')
+
     post = Post.objects.get(pk=pk)
     user = None
     for user in post.recommend.all():
         if user == request.user:
             post.recommend.remove(user)
-            break;
+            break
 
     if user != request.user:
         post.recommend.add(request.user)  # recommend: many to many field 에 현재 추천한 user 정보 입력
@@ -163,6 +179,10 @@ def recommend(request, whatboard, pk):
 
 
 def unrecommend(request, whatboard, pk):
+
+    if request.user.is_anonymous:
+        return redirect('/')
+
     post = Post.objects.get(pk=pk)
     user = None
 
@@ -181,6 +201,10 @@ def unrecommend(request, whatboard, pk):
 
 
 def deletepost(request, whatboard, pk):
+
+    if request.user.is_anonymous:
+        return redirect('/')
+
     post = Post.objects.get(pk=pk)
     if request.user == post.user:
         post.delete()
@@ -189,6 +213,10 @@ def deletepost(request, whatboard, pk):
 
 
 def deletecomment(request, whatboard, pk, commentnumber):
+
+    if request.user.is_anonymous:
+        return redirect('/')
+
     comment = Comment.objects.get(pk=commentnumber)
     if request.user == comment.user:
         comment.delete()
@@ -197,6 +225,10 @@ def deletecomment(request, whatboard, pk, commentnumber):
 
 
 def edit(request, whatboard, pk):
+
+    if request.user.is_anonymous:
+        return redirect('/')
+
     if whatboard == 'free':
         changedwhatboard = '자유게시판'
     elif whatboard == 'secret':
